requirements/base.txt
Django>=1.9,<1.10
wagtail==1.8
django-bootstrap3==7.1.0
django-ipware==1.1.5
django-pipeline==1.6.8
django-phonenumber-field==1.1.0
netaddr==0.7.18
requests==2.10.0
openpyxl==2.3.5
django-excel==0.0.4
pyexcel-ods==0.2.0
pyexcel-xls==0.2.0
pyexcel-xlsx==0.2.0
gunicorn==19.6.0
django-slack==5.7.0
django-allauth==0.29.0
wagtailtinymce==4.2.1.5
djangorestframework==3.4.0
pyxero
